+++ src/main/java/kitchenpos/response/TableGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.response;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 import java.time.LocalDateTime;
 import java.util.List;
