+++ src/main/java/kitchenpos/tablegroup/ui/TableGroupRestController.java
@@ -1,9 +1,9 @@
-package kitchenpos.ui;
+package kitchenpos.tablegroup.ui;
 
-import kitchenpos.application.TableGroupService;
-import kitchenpos.domain.TableGroup;
 import kitchenpos.request.TableGroupCreateRequest;
 import kitchenpos.response.TableGroupResponse;
+import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
