+++ src/test/java/kitchenpos/domain/TableGroupTest.java
@@ -1,13 +1,14 @@
 package kitchenpos.domain;
 
 import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-import static kitchenpos.domain.TableGroup.from;
+import static kitchenpos.tablegroup.domain.TableGroup.from;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
